Fix location

+++ src/main/java/com/entity/Address.java
@@ -16,8 +16,8 @@ public class Address {
     Long id;
     String name;
     String link;
-    Float longitude;
-    Float latitude;
+    Double longitude;
+    Double latitude;
 
     public Long getId() {
         return id;
